+++ CHANGELOG.md
@@ -6,10 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
 ### Added
-
-- New modifyer to `SKUSelector Item` for unavailable itens
+- `unavailable` modifier to `valueWrapper` CSS handle.
 
 ## [3.131.1] - 2020-10-27